Add check for leader role on store write paths (#7)

+++ server/__init__.py
@@ -1,8 +1,9 @@
-from fastapi import FastAPI, HTTPException
+from fastapi import FastAPI, HTTPException, status, Request, Response
+from fastapi.responses import JSONResponse
 from pydantic import BaseModel
 from typing import List, Union
 from store import Store
-from raft import State, Op
+from raft import State, Op, Role
 
 
 # all FastAPI and Pydantic code should be limited to the server module
@@ -32,10 +33,17 @@ class AppendEntriesResponse(BaseModel):
     success: bool
 
 
-def build_app():
+def build_app(state: State = None):
     app = FastAPI()
-    state = State()
+    if not state:
+        # dependency injection primarily for test setup and observability
+        state = State()
     state.store = Store()
+
+    # Note: Default fastapi behavior for unhandled exceptions is to return a
+    # 404. This is incorrect as a 4xx status code indicates a malformed request
+    # and should be over-ridden with a general exception middleware that
+    # returns a 500. See https://fastapi.tiangolo.com/tutorial/handling-errors/
 
     # ----- DB Router -----
     @app.get("/")
@@ -43,17 +51,29 @@ def build_app():
         """reads the specified value"""
         v = state.store.read(key)
         if not v:
-            raise HTTPException(status_code=404, detail="Not found")
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND, detail="Not found")
         return ReadResponse(value=v)
 
-    @app.post("/", status_code=201)
-    async def db_upsert(key: str, value: str):
-        """inserts the specified value without checking for previous existence"""
+    @app.post("/", status_code=status.HTTP_201_CREATED)
+    async def db_upsert(key: str, value: str, response: Response):
+        """inserts the specified value without checking for previous existence
+        """
+        if state.role != Role.LEADER:
+            return JSONResponse(
+                status_code=status.HTTP_308_PERMANENT_REDIRECT,
+                content=None,
+                headers={'Location': state.voted_for})
         state.store.upsert(key, value)
 
-    @app.delete("/", status_code=204)
+    @app.delete("/", status_code=status.HTTP_204_NO_CONTENT)
     async def db_delete(key: str):
         """deletes specified key without checking for existence"""
+        if state.role != Role.LEADER:
+            return JSONResponse(
+                status_code=status.HTTP_308_PERMANENT_REDIRECT,
+                content=None,
+                headers={'Location': state.voted_for})
         state.store.delete(key)
 
     # ----- Raft Router -----
